requirements.txt
# MiniJudge Dependencies

# Core
streamlit>=1.22.0
python-dotenv>=1.0.0

# OCR
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=9.5.0
PyMuPDF>=1.22.3

# RAG Components
langchain>=0.0.200
openai>=0.28.0
faiss-cpu>=1.7.4
numpy>=1.24.3

# LLM alternatives
transformers>=4.29.0
torch>=2.0.1
sentence-transformers>=2.2.2

# Utilities
uuid>=1.30
tqdm>=4.65.0